Fixed basis set loading and loaded raw image
In some cases (linux?), it seems that plt.imread loads the image in
rbga instead of black and white. I switched to loading the raw image.
Also, I had disabled the loading of the npy format basis sets (faster)
during debugging and forgot to enable them. Now it works again.