Tweaked the setup.py file
I made a few small changes to the setup.py file. Most importantly,
changed the author from myself to “The PyAbel Team”, to acknowledge the
hard work by all of the current (and hopefully future!) contributors to
the project.

Also, changed the description to mention that we also do the forward
Abel transform. Finally, made it clear that the lack of Cython would
only doom the C implementation of abel.direct, since we now have the
pure-Python fallback, right?