Update docs: remove misleading notation surrounding quadrant averaging
The current documentation notation describing how the symmetrize_method=="average" function works, is misleading. It says things like "Q01 = Q0 + Q1", indicating that the combination is a summation of Q0 and Q1. Instead it is either or, if only one of the quadrants is specified by use_quadrants, or it is the average of the two quadrants is specified. This commit expands the documentation to eliminate the misleading language.